+++ .gitignore
@@ -3,3 +3,4 @@
 .vscode
 results/*
 sample-media/*.mp4
+camera-simulator/rtsp_simple_server.log.txt
