+++ configs/opencv-ovms/scripts/grpc_python.py
@@ -9,24 +9,10 @@ sys.path.append("/model_server/demos/common/python")
 import argparse
 import datetime
 import cv2
-import numpy as np
 import grpc
 from client_utils import print_statistics
 from tritonclient.grpc import service_pb2, service_pb2_grpc
-from tritonclient.utils import *
-
-DataTypeToContentsFieldName = {
-    'BOOL' : 'bool_contents',
-    'BYTES' : 'bytes_contents',
-    'FP32' : 'fp32_contents',
-    'FP64' : 'fp64_contents',
-    'INT64' : 'int64_contents',
-    'INT32' : 'int_contents',
-    'UINT64' : 'uint64_contents',
-    'UINT32' : 'uint_contents',
-    'INT64' : 'int64_contents',
-    'INT32' : 'int_contents',
-}
+from grpc_postprocess import *
 
 def openInputSrc(input_src):
     # OpenCV RTSP Stream
@@ -66,43 +52,6 @@ def inference(img_str, model_name, grpc_stub):
     duration = (end_time - start_time).total_seconds() * 1000
     return [response, duration]
 
-def as_numpy(response, name):
-    index = 0
-    for output in response.outputs:
-        if output.name == name:
-            shape = []
-            for value in output.shape:
-                shape.append(value)
-            datatype = output.datatype
-            field_name = DataTypeToContentsFieldName[datatype]
-            contents = getattr(output, "contents")
-            contents = getattr(contents, f"{field_name}")
-            if index < len(response.raw_output_contents):
-                np_array = np.frombuffer(
-                    response.raw_output_contents[index], dtype=triton_to_np_dtype(output.datatype))
-            elif len(contents) != 0:
-                np_array = np.array(contents,
-                                    copy=False)
-            else:
-                np_array = np.empty(0)
-            np_array = np_array.reshape(shape)
-            return np_array
-        else:
-            index += 1
-    return None
-
-def postProcessResponse(response, duration):
-    # omz instance segmentation model has three outputs
-    output1 = as_numpy(response, "3523")
-    output2 = as_numpy(response, "3524")
-    output3 = as_numpy(response, "masks")
-    nu = np.array(output1)
-    nu2 = np.array(output2)
-    nu3 = np.array(output3)
-    # for object classification models show imagenet class
-    print('Processing time: {:.2f} ms; speed {:.2f} fps'.format(round(np.average(duration), 2),round(1000  / np.average(duration), 2)))
-    return output1
-
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Sends requests via KServe gRPC API using images in format supported by OpenCV. It displays performance statistics and optionally the model accuracy')
     parser.add_argument('--input_src', required=True, default='', help='input source for the inference pipeline')
@@ -129,4 +78,4 @@ if __name__ == '__main__':
         img_str = cv2.imencode('.jpg', img)[1].tobytes()
 
         response = inference(img_str, args['model_name'], grpc_stub)
-        postProcessResponse(response[0], response[1])
+        postProcessMaskRCNN(response[0], response[1])
