+++ configs/opencv-ovms/scripts/run_grpc_kserv.sh
@@ -13,4 +13,5 @@ fi
 
 # /scripts is mounted during the docker run 
 # python3 /scripts/grpc_infer_binary_maskrcnn-omz.py --images_list /images/inputimages.txt --grpc_address 127.0.0.1 --grpc_port $GRPC_PORT --input_name image  --batchsize $BATCH_SIZE --model_name  instance-segmentation-security-1040 2>&1
-python3 /scripts/run_grpc_python.py --input_src 'rtsp://127.0.0.1:8554/camera_0' --grpc_address 127.0.0.1 --grpc_port $GRPC_PORT --model_name  instance-segmentation-security-1040 2>&1
+# python3 /scripts/grpc_infer_binary_bit.py --images_list /images/inputimages.txt --grpc_address 127.0.0.1 --grpc_port $GRPC_PORT --input_name input_1  --batchsize $BATCH_SIZE --model_name  bit_64
+python3 /scripts/grpc_python.py --input_src 'rtsp://127.0.0.1:8554/camera_0' --grpc_address 127.0.0.1 --grpc_port $GRPC_PORT --model_name  instance-segmentation-security-1040 2>&1
